requirements.txt
blinker==1.8.2
Bottleneck==1.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
ftfy==6.3.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numexpr==2.10.1
numpy==2.0.2
openai-clip==1.0.1
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
retrying==1.3.4
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.21.0
